requirements.txt
pandas==1.5.3
mlflow==2.8.1
scikit-learn==1.2.2
numpy==1.23.5
notebook==6.5.2
matplotlib==3.6.2
python-box==6.0.2
pyYAML==6.0.1
tqdm==4.64.1
ensure==0.6.2
joblib==1.2.0
types-PyYAML
Flask==2.2.2
Flask-Cors==3.0.10